+++ app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Providers from "./providers";
-import Navbar from "./music/Navbar";
-import Footer from "./music/Footer";
+import Navbar from "./music/layout/navbar";
+import Footer from "./music/layout/footer";
 
 const inter = Inter({ subsets: ["latin"] });
